backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pandas==2.1.4
openpyxl==3.1.2
PyJWT==2.8.0
requests==2.31.0
supabase>=2.9.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
